requirements.txt
scikit-learn == 1.2.2
Pillow
fastapi
uvicorn
tensorflow
python-multipart
